+++ sms_2wayintegration/COPYRIGHT
