+++ sms_2wayintegration/__manifest__.py
@@ -6,7 +6,7 @@
     'category': 'Extra Tools',
     'summary': """Send sms 2 way integration.""",
     'version': '15.0.0.1',
-    'author': "Candidroot Solutions Pvt. Ltd.",
+    'author': "",
     'website': 'https://www.candidroot.com/',
     'sequence': 2,
     'description': """This module allows you to send sms using api""",
@@ -18,7 +18,6 @@
     'images' : ['static/description/banner.png'],
     'installable': True,
     'live_test_url': 'https://youtu.be/XTBuZzlva3I',
-    'price': 9.99,
     'currency': 'USD',
     'auto_install': False,
     'application': False,
