+++ sms_2wayintegration/models/sms_api.py
@@ -9,9 +9,9 @@ def iap_jsonrpc(url, method='call', params=None, timeout=15):
     import requests
     from odoo.http import request
 
-    user = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration_cr.msegat_user_name')
-    password = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration_cr.msegat_api_key')
-    originator = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration_cr.msegat_user_sender')
+    user = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration.msegat_user_name')
+    password = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration.msegat_api_key')
+    originator = request.env['ir.config_parameter'].sudo().get_param('sms_2wayintegration.msegat_user_sender')
     result=[]
     for p in params.get('messages', False):
         url = "https://svr22.synermaxx.asia/vmobile/suysing/api/sendnow.php?username=%s&password=%s&mobilenum=%s&fullmesg=%s&originator=%s" %(
