+++ sms_2wayintegration/views/res_config_settings_views.xml
